flask_registry: refactoring
