+++ tests/syntaxerror_module.py
@@ -21,5 +21,6 @@
 ## granted to it by virtue of its status as an Intergovernmental Organization
 ## or submit itself to any jurisdiction.
 
+
 def func_with_syntax_error()
     pass
